back/requirements.txt
asgiref==3.2.10
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Cython==0.29.14
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-formtools==2.2
django-import-export==2.3.0
django-otp==0.9.4
django-pandas==0.6.2
django-phonenumber-field==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
drf-yasg==1.20.0
et-xmlfile==1.1.0
filelock==3.0.12
gensim==3.8.3
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.6.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
loguru==0.5.3
MarkupPy==1.14
MarkupSafe==2.0.1
nltk==3.4.5
numpy==1.21.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
pandas==1.3.0
psycopg2==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sacremoses==0.0.45
scipy==1.7.0
six==1.16.0
smart-open==5.1.0
sqlparse==0.4.1
tablib==3.0.0
tokenizers==0.10.3
torch==1.9.0
tqdm==4.61.2
transformers==4.8.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
win32-setctime==1.0.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
